oopsify: move tests to their own files
