+++ oops/oopsify/oopsify_test.go
@@ -2,17 +2,22 @@ package main
 
 import (
 	"bytes"
+	"flag"
 	"go/ast"
 	"go/importer"
 	"go/parser"
 	"go/printer"
 	"go/token"
 	"go/types"
+	"io/ioutil"
+	"path"
 	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/fatih/astrewrite"
 	"github.com/kylelemons/godebug/diff"
+	"github.com/stretchr/testify/require"
 	"golang.org/x/tools/go/ast/astutil"
 )
 
@@ -36,440 +41,31 @@ type testCase struct {
 	output string
 }
 
-var testCases = []testCase{
-	{
-		name: "empty package",
-		input: `package foo
-`,
-		output: `package foo
-`,
-	},
-	{
-		name: "simple errors.New",
-		input: `package foo
-
-import (
-	"errors"
-)
-
-func f() (interface{}, error) {
-	return nil, errors.New("hello!!")
-}
-`,
-		output: `package foo
-
-import (
-	"github.com/samsarahq/go/oops"
-)
-
-func f() (interface{}, error) {
-	return nil, oops.Errorf("hello!!")
-}
-`,
-	},
-	{
-		name: "errors.New escape format",
-		input: `package foo
-
-import (
-	"errors"
-)
-
-func f() (interface{}, error) {
-	return nil, errors.New("hello!! %s")
-}
-`,
-		output: `package foo
-
-import (
-	"github.com/samsarahq/go/oops"
-)
-
-func f() (interface{}, error) {
-	return nil, oops.Errorf("hello!! %%s")
-}
-`,
-	},
-	{
-		name: "errors.New complex",
-		input: `package foo
-
-import (
-	"errors"
-)
-
-func g() string {
-	return ""
-}
-
-func f() (interface{}, error) {
-	return nil, errors.New(g())
-}
-`,
-		output: `package foo
-
-import (
-	"github.com/samsarahq/go/oops"
-)
-
-func g() string {
-	return ""
-}
-
-func f() (interface{}, error) {
-	return nil, oops.Errorf("%s", g())
-}
-`,
-	},
-	{
-		name: "simple fmt.Errorf",
-		input: `package foo
-
-import (
-	"fmt"
-)
-
-func f() (interface{}, error) {
-	return nil, fmt.Errorf("some thing went wrong %s %d", "foo", 10)
-}
-`,
-		output: `package foo
-
-import (
-	"github.com/samsarahq/go/oops"
-)
-
-func f() (interface{}, error) {
-	return nil, oops.Errorf("some thing went wrong %s %d", "foo", 10)
-}
-`,
-	},
-	{
-		name: "fmt.Errorf with %s error",
-		input: `package foo
-
-import (
-	"fmt"
-)
-
-func f() (interface{}, error) {
-	var err error
-	return nil, fmt.Errorf("some thing went wrong: %s", err)
-}
-`,
-		output: `package foo
-
-import (
-	"github.com/samsarahq/go/oops"
-)
-
-func f() (interface{}, error) {
-	var err error
-	return nil, oops.Wrapf(err, "some thing went wrong")
-}
-`,
-	},
-	{
-		name: "wrap global error",
-		input: `package foo
-
-import (
-	"errors"
-)
-
-var ErrBad = errors.New("bad")
-
-func f() (interface{}, error) {
-	return nil, ErrBad
-}
-`,
-		output: `package foo
-
-import (
-	"errors"
-	"github.com/samsarahq/go/oops"
-)
-
-var ErrBad = errors.New("bad")
-
-func f() (interface{}, error) {
-	return nil, oops.Wrapf(ErrBad, "")
-}
-`,
-	},
-	{
-		name: "wrap custom error type",
-		input: `package foo
-
-type SnazzyError struct {
-}
-
-func NewSnazzyError() *SnazzyError {
-	// Should not be wrapped as error is not assignale to *SnazzyError.
-	return &SnazzyError{}
-}
-
-func (s *SnazzyError) Error() string {
-	return "I am quite snazzy"
-}
-
-func f() (interface{}, error) {
-	return nil, NewSnazzyError()
-}
-`,
-		output: `package foo
-
-import "github.com/samsarahq/go/oops"
-
-type SnazzyError struct {
-}
-
-func NewSnazzyError() *SnazzyError {
-	// Should not be wrapped as error is not assignale to *SnazzyError.
-	return &SnazzyError{}
-}
-
-func (s *SnazzyError) Error() string {
-	return "I am quite snazzy"
-}
-
-func f() (interface{}, error) {
-	return nil, oops.Wrapf(NewSnazzyError(), "")
-}
-`,
-	},
-	{
-		name: "check nil",
-		input: `package foo
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err == nil {
-	}
-}
-`,
-		output: `package foo
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err == nil {
-	}
-}
-`,
-	},
-	{
-		name: "check non-nil",
-		input: `package foo
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != nil {
-	}
-}
-`,
-		output: `package foo
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != nil {
-	}
-}
-`,
-	},
-	{
-		name: "check well known",
-		input: `package foo
-
-import (
-	"io"
-)
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != io.EOF {
-	}
-}
-`,
-		output: `package foo
-
-import (
-	"io"
-	"github.com/samsarahq/go/oops"
-)
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); oops.Cause(err) != io.EOF {
-	}
-}
-`,
-	},
-	{
-		name: "error strings.Contains",
-		input: `package foo
-
-import (
-	"strings"
-)
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != nil && !strings.Contains(err.Error(), "some string") {
-	}
-}
-`,
-		output: `package foo
-
-import (
-	"strings"
-	"github.com/samsarahq/go/oops"
-)
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != nil && !strings.Contains(oops.Cause(err).Error(), "some string") {
-	}
-}
-`,
-	},
-	{
-		name: "normal strings.Contains",
-		input: `package foo
-
-import (
-	"strings"
-)
-
-func f() bool {
-	return strings.Contains("foo", "bar")
-}
-`,
-		output: `package foo
-
-import (
-	"strings"
-)
-
-func f() bool {
-	return strings.Contains("foo", "bar")
-}
-`,
-	},
-	{
-		name: "assert custom error type",
-		input: `package foo
-
-type SnazzyError struct {
-}
-
-func (s *SnazzyError) Error() string {
-	return "I am quite snazzy"
-}
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != nil {
-		if _, ok := err.(*SnazzyError); ok {
+func loadTestCases(t *testing.T) []testCase {
+	files, err := ioutil.ReadDir("testdata/")
+	require.NoError(t, err)
+
+	var testCases []testCase
+
+	for _, file := range files {
+		if strings.HasSuffix(file.Name(), ".in.go") {
+			name := strings.TrimSuffix(file.Name(), ".in.go")
+			inputName := file.Name()
+			outputName := name + ".out.go"
+
+			inputBytes, err := ioutil.ReadFile(path.Join("testdata", inputName))
+			require.NoError(t, err)
+			outputBytes, _ := ioutil.ReadFile(path.Join("testdata", outputName))
+
+			testCases = append(testCases, testCase{
+				name:   name,
+				input:  string(inputBytes),
+				output: string(outputBytes),
+			})
 		}
 	}
-}
-`,
-		output: `package foo
 
-import "github.com/samsarahq/go/oops"
-
-type SnazzyError struct {
-}
-
-func (s *SnazzyError) Error() string {
-	return "I am quite snazzy"
-}
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != nil {
-		if _, ok := oops.Cause(err).(*SnazzyError); ok {
-		}
-	}
-}
-`,
-	},
-	{
-		name: "switch custom error type",
-		input: `package foo
-
-type SnazzyError struct {
-}
-
-func (s *SnazzyError) Error() string {
-	return "I am quite snazzy"
-}
-
-func f() error {
-	return nil
-}
-
-
-func g() {
-	if err := f(); err != nil {
-		switch err.(type) {
-		}
-	}
-}
-`,
-		output: `package foo
-
-import "github.com/samsarahq/go/oops"
-
-type SnazzyError struct {
-}
-
-func (s *SnazzyError) Error() string {
-	return "I am quite snazzy"
-}
-
-func f() error {
-	return nil
-}
-
-func g() {
-	if err := f(); err != nil {
-		switch oops.Cause(err).(type) {
-		}
-	}
-}
-`,
-	},
+	return testCases
 }
 
 func do(t *testing.T, source string) string {
@@ -512,11 +108,16 @@ func do(t *testing.T, source string) string {
 	return buffer.String()
 }
 
+var rewriteSnapshots = flag.Bool("rewriteSnapshots", false, "rewrite .out.go files")
+
 func TestOopsify(t *testing.T) {
-	for _, testCase := range testCases {
+	for _, testCase := range loadTestCases(t) {
 		t.Run(testCase.name, func(t *testing.T) {
 			oopsified := do(t, testCase.input)
-			if diff := diff.Diff(oopsified, testCase.output); diff != "" {
+
+			if *rewriteSnapshots {
+				ioutil.WriteFile(path.Join("testdata", testCase.name+".out.go"), []byte(oopsified), 0644)
+			} else if diff := diff.Diff(oopsified, testCase.output); diff != "" {
 				t.Error(diff)
 			}
 		})
@@ -524,7 +125,7 @@ func TestOopsify(t *testing.T) {
 }
 
 func TestOopsifyIdempotent(t *testing.T) {
-	for _, testCase := range testCases {
+	for _, testCase := range loadTestCases(t) {
 		t.Run(testCase.name, func(t *testing.T) {
 			once := do(t, testCase.input)
 			twice := do(t, once)
