+++ oops/oopsify/testdata/check_non_nil.in.go
@@ -1,0 +1,10 @@
+package foo
+
+func f() error {
+	return nil
+}
+
+func g() {
+	if err := f(); err != nil {
+	}
+}
