+++ oops/oopsify/testdata/check_well_known.in.go
@@ -1,0 +1,14 @@
+package foo
+
+import (
+	"io"
+)
+
+func f() error {
+	return nil
+}
+
+func g() {
+	if err := f(); err != io.EOF {
+	}
+}
