+++ oops/oopsify/testdata/check_well_known.out.go
@@ -1,0 +1,15 @@
+package foo
+
+import (
+	"io"
+	"github.com/samsarahq/go/oops"
+)
+
+func f() error {
+	return nil
+}
+
+func g() {
+	if err := f(); oops.Cause(err) != io.EOF {
+	}
+}
