+++ oops/oopsify/testdata/empty.in.go
@@ -1,0 +1,1 @@
+package foo
