+++ oops/oopsify/testdata/empty.out.go
@@ -1,0 +1,1 @@
+package foo
