+++ oops/oopsify/testdata/errors_new_complex.in.go
@@ -1,0 +1,13 @@
+package foo
+
+import (
+	"errors"
+)
+
+func g() string {
+	return ""
+}
+
+func f() (interface{}, error) {
+	return nil, errors.New(g())
+}
