+++ oops/oopsify/testdata/errors_new_complex.out.go
@@ -1,0 +1,13 @@
+package foo
+
+import (
+	"github.com/samsarahq/go/oops"
+)
+
+func g() string {
+	return ""
+}
+
+func f() (interface{}, error) {
+	return nil, oops.Errorf("%s", g())
+}
