+++ oops/oopsify/testdata/errors_new_escape_format.in.go
@@ -1,0 +1,9 @@
+package foo
+
+import (
+	"errors"
+)
+
+func f() (interface{}, error) {
+	return nil, errors.New("hello!! %s")
+}
