+++ oops/oopsify/testdata/errors_new_simple.out.go
@@ -1,0 +1,9 @@
+package foo
+
+import (
+	"github.com/samsarahq/go/oops"
+)
+
+func f() (interface{}, error) {
+	return nil, oops.Errorf("hello!!")
+}
