+++ oops/oopsify/testdata/fmt_errorf_simple.in.go
@@ -1,0 +1,9 @@
+package foo
+
+import (
+	"fmt"
+)
+
+func f() (interface{}, error) {
+	return nil, fmt.Errorf("some thing went wrong %s %d", "foo", 10)
+}
