+++ oops/oopsify/testdata/fmt_errorf_with_err.in.go
@@ -1,0 +1,10 @@
+package foo
+
+import (
+	"fmt"
+)
+
+func f() (interface{}, error) {
+	var err error
+	return nil, fmt.Errorf("some thing went wrong: %s", err)
+}
