+++ oops/oopsify/testdata/fmt_errorf_with_err.out.go
@@ -1,0 +1,10 @@
+package foo
+
+import (
+	"github.com/samsarahq/go/oops"
+)
+
+func f() (interface{}, error) {
+	var err error
+	return nil, oops.Wrapf(err, "some thing went wrong")
+}
