+++ oops/oopsify/testdata/strings_contains_error.in.go
@@ -1,0 +1,14 @@
+package foo
+
+import (
+	"strings"
+)
+
+func f() error {
+	return nil
+}
+
+func g() {
+	if err := f(); err != nil && !strings.Contains(err.Error(), "some string") {
+	}
+}
