+++ oops/oopsify/testdata/strings_contains_error.out.go
@@ -1,0 +1,15 @@
+package foo
+
+import (
+	"strings"
+	"github.com/samsarahq/go/oops"
+)
+
+func f() error {
+	return nil
+}
+
+func g() {
+	if err := f(); err != nil && !strings.Contains(oops.Cause(err).Error(), "some string") {
+	}
+}
