+++ oops/oopsify/testdata/strings_contains_plain.out.go
@@ -1,0 +1,9 @@
+package foo
+
+import (
+	"strings"
+)
+
+func f() bool {
+	return strings.Contains("foo", "bar")
+}
