+++ oops/oopsify/testdata/well_known_err.in.go
@@ -1,0 +1,11 @@
+package foo
+
+import (
+	"errors"
+)
+
+var ErrBad = errors.New("bad")
+
+func f() (interface{}, error) {
+	return nil, ErrBad
+}
