+++ oops/oopsify/testdata/well_known_err.out.go
@@ -1,0 +1,12 @@
+package foo
+
+import (
+	"errors"
+	"github.com/samsarahq/go/oops"
+)
+
+var ErrBad = errors.New("bad")
+
+func f() (interface{}, error) {
+	return nil, oops.Wrapf(ErrBad, "")
+}
